fix: satisfy pkg/api/contracttests file-count limit after durable test split
